1. DuplicatedIdException.java, InvalidIdException.java를 common 패키지로 이동 2. DormitoryNotFoundException.java 생성 및 ErrorCode, ControllerExceptionHandler.java에 추가 3. DormitoryService.java에서 중복 체크 method 이름 변경 (duplicateCheck -> isExist) 4. DormitoryService.java에서 수정, 삭제 method에 DormitoryNotFoundException 예외 추가 5. DormitoryService 테스트 코드 작성 6. LoginControllerTest.java 자잘한 수정

+++ src/main/java/com/project/clickit/service/DormitoryService.java
@@ -2,6 +2,8 @@ package com.project.clickit.service;
 
 import com.project.clickit.dto.DormitoryDTO;
 import com.project.clickit.entity.DormitoryEntity;
+import com.project.clickit.exceptions.common.DuplicatedIdException;
+import com.project.clickit.exceptions.dormitory.DormitoryNotFoundException;
 import com.project.clickit.repository.DormitoryRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -21,12 +23,15 @@ public class DormitoryService {
     }
 
     @Transactional
-    public Boolean duplicateCheck(String id) {
+    public Boolean isExist(String id) {
         return dormitoryRepository.existsById(id);
     }
 
     @Transactional
     public void createDormitory(DormitoryDTO dormitoryDTO) {
+        if(isExist(dormitoryDTO.getId())){
+            throw new DuplicatedIdException();
+        }
         dormitoryRepository.save(dormitoryDTO.toEntity());
     }
 
@@ -42,11 +47,17 @@ public class DormitoryService {
 
     @Transactional
     public void updateDormitoryName(String id, String name) {
+        if(!isExist(id)){
+            throw new DormitoryNotFoundException();
+        }
         dormitoryRepository.updateDormitoryName(id, name);
     }
 
     @Transactional
     public void deleteById(String id) {
+        if(!isExist(id)){
+            throw new DormitoryNotFoundException();
+        }
         dormitoryRepository.deleteById(id);
     }
